refactoting grammar
