+++ src/atomizer.js
@@ -7,117 +7,13 @@
 'use strict';
 
 var _ = require('lodash');
-var utils = require('./utils');
-var JSS = require('./jss');
+var utils = require('./lib/utils');
+var JSS = require('./lib/jss');
+var Grammar = require('./lib/grammar');
 var objectAssign = require('object-assign');
 var XRegExp = require('xregexp').XRegExp;
 
 var RULES = require('./rules.js').concat(require('./helpers.js'));
-
-var PSEUDOS = {
-    ':active':          ':a',
-    ':checked':         ':c',
-    ':default':         ':d',
-    ':disabled':        ':di',
-    ':empty':           ':e',
-    ':enabled':         ':en',
-    ':first':           ':fi',
-    ':first-child':     ':fc',
-    ':first-of-type':   ':fot',
-    ':fullscreen':      ':fs',
-    ':focus':           ':f',
-    ':hover':           ':h',
-    ':indeterminate':   ':ind',
-    ':in-range':        ':ir',
-    ':invalid':         ':inv',
-    ':last-child':      ':lc',
-    ':last-of-type':    ':lot',
-    ':left':            ':l',
-    ':link':            ':li',
-    ':only-child':      ':oc',
-    ':only-of-type':    ':oot',
-    ':optional':        ':o',
-    ':out-of-range':    ':oor',
-    ':read-only':       ':ro',
-    ':read-write':      ':rw',
-    ':required':        ':req',
-    ':right':           ':r',
-    ':root':            ':rt',
-    ':scope':           ':s',
-    ':target':          ':t',
-    ':valid':           ':va',
-    ':visited':         ':vi'
-};
-var PSEUDOS_INVERTED = _.invert(PSEUDOS);
-var PSEUDO_REGEX = [];
-for (var pseudo in PSEUDOS) {
-    PSEUDO_REGEX.push(pseudo);
-    PSEUDO_REGEX.push(PSEUDOS[pseudo]);
-}
-PSEUDO_REGEX = '(?:' + PSEUDO_REGEX.join('|') + ')(?![a-z])';
-
-// regular grammar to match valid atomic classes
-var GRAMMAR = {
-    'BOUNDARY'   : '(?:^|\\s|"|\'|\{)',
-    'PARENT'     : '[a-zA-Z][-_a-zA-Z0-9]+?',
-    'PARENT_SEP' : '[>_+]',
-    // all character allowed to be in values
-    'VALUES'     : '[-_,.#$/%0-9a-zA-Z]+',
-    'FRACTION'   : '(?<numerator>[0-9]+)\\/(?<denominator>[1-9](?:[0-9]+)?)',
-    'PARAMS'     : '\\((?<params>[^)]*)\\)',
-    'NUMBER'     : '-?[0-9]+(?:\.[0-9]+)?|\\.[0-9]+',
-    'UNIT'       : '[a-zA-Z%]+',
-    'HEX'        : '#[0-9a-f]{3}(?:[0-9a-f]{3})?',
-    'ALPHA'      : '\\.\\d{1,2}',
-    'IMPORTANT'  : '!',
-    // https://regex101.com/r/mM2vT9/8
-    'NAMED'      : '([\\w$]+(?:(?:-(?!\\-))?\\w*)*)',
-    'PSEUDO'     : PSEUDO_REGEX,
-    'BREAKPOINT' : '--(?<breakPoint>[a-z]+)'
-};
-GRAMMAR.PARENT_SELECTOR = [
-    // parent (any character that is not a space)
-    '(?<parent>',
-        GRAMMAR.PARENT,
-    ')',
-    // followed by optional pseudo class
-    '(?<parentPseudo>',
-        GRAMMAR.PSEUDO,
-    ')?',
-    // followed by either a descendant or direct symbol
-    '(?<parentSep>',
-        GRAMMAR.PARENT_SEP,
-    ')'
-].join('');
-
-var VALUE_SYNTAXE = XRegExp([
-    '(?<fraction>',
-        GRAMMAR.FRACTION,
-    ')',
-    '|',
-    '(?:',
-        '(?<hex>',
-            GRAMMAR.HEX,
-        ')',
-        '(?<alpha>',
-            GRAMMAR.ALPHA,
-        ')?',
-        '(?!',
-            GRAMMAR.UNIT,
-        ')',
-    ')',
-    '|',
-    '(?<number>',
-        GRAMMAR.NUMBER,
-    ')',
-    '(?<unit>',
-        GRAMMAR.UNIT,
-    ')?',
-    '|',
-    '(?<named>',
-        GRAMMAR.NAMED,
-    ')',
-].join(''));
 
 /**
  * constructor
@@ -137,7 +33,6 @@ function Atomizer(options/*:AtomizerOptions*/, rules/*:AtomizerRules*/) {
  * @public
  */
 Atomizer.prototype.addRules = function(rules/*:AtomizerRules*/)/*:void*/ {
-
     rules.forEach(function (rule) {
         if (
             (rule.type === 'pattern' && this.rulesMap.hasOwnProperty(rule.matcher)) ||
@@ -165,88 +60,12 @@ Atomizer.prototype.addRules = function(rules/*:AtomizerRules*/)/*:void*/ {
  * @private
  */
 Atomizer.prototype.getSyntax = function ()/*:string*/ {
-    var syntax;
-    var helperRegex;
-    var propRegex;
-    var helpersKeys;
-    var rulesKeys;
-    var mainSyntax;
-
-    if (this.syntax) {
-        return this.syntax;
-    } else {
-        // sort matchers by descending alphabetical order
-        // this is important so "B" doesn't match "Bgc"
-        // e.g. Use (Bgc|B) instead of (B|Bgc)
-        helpersKeys = Object.keys(this.helpersMap).sort(function (a, b) {
-            return a > b ? -1 : 1;
-        }).join('|');
-        rulesKeys = Object.keys(this.rulesMap).sort(function (a, b) {
-            return a > b ? -1 : 1;
-        }).join('|');
-
-        // helpers regex
-        if (helpersKeys.length) {
-            helperRegex = [
-                // matcher
-                '(?<helper>',
-                    helpersKeys,
-                ')',
-                // value is optional
-                '(?:\\(',
-                    '(?<helperValues>',
-                        GRAMMAR.VALUES,
-                    ')',
-                '\\))?',
-            ].join('');
-            mainSyntax = helperRegex;
-        }
-        // rules regex
-        if (rulesKeys.length) {
-            propRegex = [
-                // matcher
-                '(?<prop>',
-                    rulesKeys,
-                ')',
-                // value is required
-                '(?:\\(',
-                    '(?<atomicValues>',
-                        GRAMMAR.VALUES,
-                    ')',
-                '\\))',
-            ].join('');
-            mainSyntax = propRegex;
-        }
-
-        if (helpersKeys.length && rulesKeys.length) {
-            mainSyntax = ['(?:', propRegex, '|', helperRegex,')'].join('');
-        }
-
-        syntax = [
-            // word boundary
-            GRAMMAR.BOUNDARY,
-            // optional parent
-            '(?<parentSelector>',
-                GRAMMAR.PARENT_SELECTOR,
-            ')?',
-            mainSyntax,
-            '(?<important>',
-                GRAMMAR.IMPORTANT,
-            ')?',
-            // optional pseudo
-            '(?<valuePseudo>',
-                GRAMMAR.PSEUDO,
-            ')?',
-            // optional modifier
-            '(?:',
-                GRAMMAR.BREAKPOINT,
-            ')?'
-        ].join('');
-
-        this.syntax = XRegExp(syntax, 'g');
-
-        return this.syntax;
+    if (!this.syntax) {
+       // All Grammar and syntax parsing  should be in the Grammar class
+       this.syntax =  new Grammar(this.rulesMap, this.helpersMap).getSyntax();
     }
+
+    return this.syntax;
 };
 
 /**
@@ -264,7 +83,8 @@ Atomizer.prototype.findClassNames = function (src/*:string*/)/*:string[]*/ {
         className = match[0].substr(1);
 
         // assign to classNamesObj as key and give it a counter
-        classNamesObj[className] = classNamesObj[className] ? classNamesObj[className] + 1 : 1;
+        classNamesObj[className] = (classNamesObj[className] || 0) + 1;
+
         // run regex again
         match = classNameSyntax.exec(src);
     }
@@ -362,7 +182,7 @@ Atomizer.prototype.parseConfig = function (config/*:AtomizerConfig*/, options/*:
             // values can be separated by a comma
             // parse them and return a valid value
             values = values.split(',').map(function (value, index) {
-                var matchVal = XRegExp.exec(value, VALUE_SYNTAXE);
+                var matchVal = Grammar.matchValue(value);
                 var propAndValue;
 
                 if (matchVal.number) {
@@ -642,7 +462,7 @@ Atomizer.prototype.getCss = function (config/*:AtomizerConfig*/, options/*:CSSOp
  * get non abbreviated pseudo class string given abbreviated or non abbreviated form
  */
 Atomizer.getPseudo = function (pseudoName/*:string*/)/*:string*/ {
-    return PSEUDOS[pseudoName] ? pseudoName : PSEUDOS_INVERTED[pseudoName];
+    return Grammar.getPseudo(pseudoName);
 };
 
 /**
